+++ tv/src/main/java/com/abrenoch/hyperiongrabber/tv/fragments/settings/BasicSettingsStepFragment.kt
@@ -1,12 +1,22 @@
 package com.abrenoch.hyperiongrabber.tv.fragments.settings
 
 import android.app.Activity
+import android.content.Context
+import android.graphics.Color
+import android.os.AsyncTask
 import android.os.Bundle
 import android.support.v17.leanback.widget.GuidanceStylist
 import android.support.v17.leanback.widget.GuidedAction
+import android.widget.Toast
+import com.abrenoch.hyperiongrabber.common.network.Hyperion
 import com.abrenoch.hyperiongrabber.tv.R
+import java.lang.ref.WeakReference
+import java.net.UnknownHostException
 
 internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
+
+    /** The amount of times the connection was tested */
+    private var testCounter = 0
 
     override fun onProvideTheme(): Int {
         return R.style.Theme_HyperionGrabber_GuidedStep_First
@@ -32,6 +42,18 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
                 ACTION_PORT,
                 getString(R.string.pref_title_port),
                 prefs.getInt(R.string.pref_key_port).toString()
+        )
+
+        val enterHorizontalLEDCount = unSignedNumberAction(
+                ACTION_X_LED_COUNT,
+                getString(R.string.pref_title_x_led),
+                prefs.getInt(R.string.pref_key_x_led).toString()
+        )
+
+        val enterVerticalLEDCount = unSignedNumberAction(
+                ACTION_Y_LED_COUNT,
+                getString(R.string.pref_title_y_led),
+                prefs.getInt(R.string.pref_key_y_led).toString()
         )
 
         val startOnBootEnabled = prefs.getBoolean(R.string.pref_key_boot)
@@ -158,6 +180,8 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
 
         actions.add(enterHost)
         actions.add(enterPort)
+        actions.add(enterHorizontalLEDCount)
+        actions.add(enterVerticalLEDCount)
         actions.add(startOnBoot)
         actions.add(advancedInfo)
         actions.add(priority)
@@ -169,6 +193,13 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
 
     override fun onCreateButtonActions(actions: MutableList<GuidedAction>, savedInstanceState: Bundle?) {
         actions.add(continueAction())
+        actions.add(GuidedAction.Builder(context)
+                .id(ACTION_TEST)
+                .title(getString(R.string.guidedstep_test))
+                .description(R.string.guidedstep_test_description)
+                .build()
+
+        )
         actions.add(backAction())
     }
 
@@ -178,7 +209,10 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
             try {
                 val host = assertStringValue(ACTION_HOST_NAME)
                 val port = assertIntValue(ACTION_PORT)
+                val xLED = assertIntValue(ACTION_X_LED_COUNT)
+                val yLED = assertIntValue(ACTION_Y_LED_COUNT)
                 val startOnBootEnabled = findActionById(ACTION_START_ON_BOOT).isChecked
+                val priority = assertIntValue(ACTION_MESSAGE_PRIORITY)
                 val frameRate = assertSubActionValue(ACTION_CAPTURE_RATE, String::class.java)
                 val useOgl = assertSubActionValue(ACTION_GRABBER_GROUP, Boolean::class.java)
                 val reconnect = findSubActionById(ACTION_RECONNECT)!!.isChecked
@@ -186,7 +220,10 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
 
                 prefs.putString(R.string.pref_key_host, host)
                 prefs.putInt(R.string.pref_key_port, port)
+                prefs.putInt(R.string.pref_key_x_led, xLED)
+                prefs.putInt(R.string.pref_key_y_led, yLED)
                 prefs.putBoolean(R.string.pref_key_boot, startOnBootEnabled)
+                prefs.putInt(R.string.pref_key_priority, priority)
                 prefs.putInt(R.string.pref_key_reconnect_delay, reconnectDelay)
                 prefs.putString(R.string.pref_key_framerate, frameRate)
                 prefs.putBoolean(R.string.pref_key_reconnect, reconnect)
@@ -201,6 +238,15 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
 
             return
 
+        } else if (action.id == ACTION_TEST){
+            val colorIdx = testCounter % TEST_COLORS.size
+            val color = TEST_COLORS[colorIdx]
+            testCounter++
+
+            val host = assertStringValue(ACTION_HOST_NAME)
+            val port = assertIntValue(ACTION_PORT)
+            val priority = assertIntValue(ACTION_MESSAGE_PRIORITY)
+            testHyperionColor(host, port, priority, color)
         }
 
         super.onGuidedActionClicked(action)
@@ -240,10 +286,17 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
         return super.onSubGuidedActionClicked(action)
     }
 
+    /** tries to connect to Hyperion and sets the given color for 5 seconds  */
+    private fun testHyperionColor(hostName: String, port: Int, priority: Int, color: Int) {
+        TestTask(context).execute(TestSpec(hostName, port, priority, color))
+    }
+
     companion object {
         private const val ACTION_HOST_NAME = 100L
         private const val ACTION_PORT = 110L
         private const val ACTION_START_ON_BOOT = 120L
+        private const val ACTION_X_LED_COUNT = 130L
+        private const val ACTION_Y_LED_COUNT = 140L
         private const val ACTION_RECONNECT_GROUP = 200L
         private const val ACTION_RECONNECT = 210L
         private const val ACTION_RECONNECT_DELAY = 220L
@@ -254,6 +307,58 @@ internal class BasicSettingsStepFragment : SettingsStepBaseFragment() {
         private const val ACTION_GRABBER_SET_ID = 550
         private const val ACTION_GRABBER_MEDIA = 560L
         private const val ACTION_GRABBER_OGL = 570L
+        private const val ACTION_TEST = 600L
+
+        private val TEST_COLORS = intArrayOf(Color.RED, Color.GREEN, Color.BLUE, Color.WHITE)
+
+        class TestTask(context: Context): AsyncTask<TestSpec, Int, Int>(){
+
+
+            private val contextRef = WeakReference(context)
+
+            override fun doInBackground(vararg params: TestSpec?): Int {
+                try {
+                    val (host, port, priority, color) = params[0]!!
+                    val hyperion = Hyperion(host, port)
+                    if (hyperion.isConnected) {
+                        hyperion.setColor(color, priority, 3000)
+                    } else {
+                        return RESULT_UNREACHABLE
+                    }
+                    hyperion.disconnect()
+                    return RESULT_OK
+                } catch (e: UnknownHostException) {
+                    return RESULT_UNREACHABLE
+                } catch (e: Exception){
+                    return RESULT_UNKNOWN_ERROR
+                }
+            }
+
+            override fun onPostExecute(result: Int) {
+                contextRef.get()?.run {
+                    val messageRes = when(result){
+                        RESULT_OK -> R.string.guidedstep_test_success
+                        RESULT_UNREACHABLE -> R.string.guidedstep_test_host_unreachable
+                        else -> R.string.guidedstep_test_unknown_error
+                    }
+
+                    Toast.makeText(this, messageRes, Toast.LENGTH_SHORT).show()
+                }
+
+
+
+            }
+
+            companion object {
+                const val RESULT_OK = 1
+                const val RESULT_UNREACHABLE = -1
+                const val RESULT_UNKNOWN_ERROR = -100
+            }
+
+        }
+
+        data class TestSpec(val host: String, val port: Int, val priority: Int, val color: Int)
+
     }