+++ CHANGELOG.md
@@ -6,11 +6,14 @@
 - Reconnect behavior implemented for mobile build
 - New connection wizard
 - New settings/connection page (tv build)
+- Quick settings tile to toggle grabber (mobile build)
 
 ### Fixed
 - Grabber would fail to resume when waking device
 - OpenGL grabber sometimes halting immediately after starting screen grab
 - Default grabber failing to send data the first time it is turned on
+- Grabber not stoping when the host is unreachable
+- Aspect ratio of grabbed image being slightly off
 
 ## [v0.3-alpha]
 ### Changes